Refactor JpegColorConverters
